+++ src/psyclone/nemo.py
@@ -47,12 +47,13 @@ from fparser.two import Fortran2003
 from psyclone.configuration import Config
 from psyclone.psyGen import PSy, Invokes, Invoke, InvokeSchedule, Node, \
     Loop, CodedKern, InternalError, NameSpaceFactory, Schedule, \
-    Fparser2ASTProcessor, colored, SCHEDULE_COLOUR_MAP
+    colored, SCHEDULE_COLOUR_MAP
+from psyclone.psyir.frontend.fparser2 import Fparser2Reader
 
 
-class NemoFparser2ASTProcessor(Fparser2ASTProcessor):
+class NemoFparser2Reader(Fparser2Reader):
     '''
-    Specialisation of Fparser2ASTProcessor for the Nemo API.
+    Specialisation of Fparser2Reader for the Nemo API.
     '''
     def _create_loop(self, parent, variable_name):
         '''
@@ -126,7 +127,7 @@ class NemoFparser2ASTProcessor(Fparser2ASTProcessor):
         '''
         if NemoImplicitLoop.match(child):
             return NemoImplicitLoop(child, parent=parent)
-        return super(NemoFparser2ASTProcessor,
+        return super(NemoFparser2Reader,
                      self)._create_child(child, parent=parent)
 
 
@@ -284,7 +285,7 @@ class NemoPSy(PSy):
         return self._ast
 
 
-class NemoInvokeSchedule(InvokeSchedule, NemoFparser2ASTProcessor):
+class NemoInvokeSchedule(InvokeSchedule, NemoFparser2Reader):
     '''
     The NEMO-specific InvokeSchedule sub-class. This is the top-level node in
     PSyclone's IR of a NEMO program unit (program, subroutine etc).
@@ -301,7 +302,7 @@ class NemoInvokeSchedule(InvokeSchedule, NemoFparser2ASTProcessor):
     def __init__(self, invoke, ast):
         # pylint: disable=super-init-not-called, non-parent-init-called
         Node.__init__(self)
-        NemoFparser2ASTProcessor.__init__(self)
+        NemoFparser2Reader.__init__(self)
 
         self._invoke = invoke
         self._ast = ast
