+++ src/psyclone/transformations.py
@@ -3475,7 +3475,7 @@ class NemoExplicitLoopTrans(Transformation):
         psyir_parent.children.remove(loop)
         # Next, we simply process the transformed fparser2 AST to generate
         # the new PSyIR of it
-        astprocessor = nemo.NemoFparser2ASTProcessor()
+        astprocessor = nemo.NemoFparser2Reader()
         astprocessor.process_nodes(psyir_parent, [new_loop], loop.ast._parent)
         # Delete the old PSyIR node that we have transformed
         del loop
